backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
flask-cors==4.0.0
psycopg2-binary==2.9.11
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
